clean and fix code and notebook - code runs
